Prepare requests for checkout (#98)
* Don't overwrite body with query data

* isPrepared checks authorization and Kickstarter-iOS-App header

Sometimes authorization can be nil (e.g. if user is not logged in,
and there is no basic http auth) – so the old check was not
sufficient.

I've also noticed sometimes the Kickstarter-iOS-App header can
be set, but no token (e.g. with redirects around login), so
I think we at least need to check both these headers.

* Update Prelude

* Update test, should no longer set body with header data

* Create adhoc service with basic auth

* Add tests for isPrepared

Covers the situation where there is no oauth token, and also no
basic auth

* Overwrite body if not already set

This is something of a hack to allow update drafts to work. We'll
probably have to address this properly at some stage.